smart_contract.sol
Come and fund for a particular cause and have voting right in the usage of these funds.

All NGOs use websites on web2.0 to take funds from the common masses and use them for charity purposes. But many times all this money has not been used the way it should have been used. One of the causes for this is the website being hacked by the party not the authority and the hacker can use all the funding as per his / her greed. To solve this problem I have used the feature of only allowing the manager of the project to deploy this smart contract. The second problem is that not everyone can fund to these kinds of projects as they require KYC information to be submitted and this smart contract solves this project as web3.0 is open to anyone with an internet connection and with a minimum amount of ether that is set by the manager at the time of deployment. Third and the most important problem that needs to be solved is that the funds are being used in the way the majority of the fund givers want them to be used this problem is solved by my smart contract as before using all the funds the manager needs to validate the purpose of use by more than 50% of the fund givers.
All these problems are solved through a smart contract on the blockchain so that assure that there is transparency of the funds -like the amount of funds collected, the number of fund givers, and the result of the votes and also makes it transparent that the manager sends all the funds to the address validated by more than 50 % of fund givers.